requirements.txt
absl-py==0.9.0
astunparse==1.6.3
bleach==1.5.0
boto==2.49.0
boto3==1.13.16
botocore==1.16.16
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
configparser==5.0.0
coverage==5.1
docutils==0.15.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Executor==0.9.3
flask-mongoengine==0.9.5
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
gast==0.3.3
gensim==3.8.3
google-auth==1.16.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
html5lib==0.9999999
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
jieba==0.42.1
Jinja2==2.11.2
jmespath==0.10.0
Keras-Preprocessing==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
mongoengine==0.20.0
nose==1.3.7
numpy==1.18.0
oauthlib==3.1.0
opt-einsum==3.2.1
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.10.1
python-dateutil==2.8.1
PyYAML==5.3.1
rednose==1.3.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
scipy==1.4.1
six==1.15.0
smart-open==2.0.0
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
termstyle==0.1.11
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
zipp==3.1.0
